Improve assignment validation rmd
Explicitly define the facility and area types and used them
to order the report tables logically. This does mean that to
generalize the RMD, it will need to be enhanced in the future
to accept command land arguments including these defs.

Cleaned up the table number formats (decimals and
delimeters).