Change macro to take a link_params array instead of param_file
The parameter file approach limits the flexibility of the macro.
The unit test still uses a parameter file, demonstrating that it
is still possible, but the parameter file is read into the options
array before calling the macro. This means that other methods
can be used to feed the link parameters into the macro - including
just manually writing them in the project-level code that calls the
delay allocation method.