Remove the purpose loop for MC
The generic model will instead take period as an input.
This allows it to be agnostic to how each individual model
handles time of day. In traditional models, the project code
will just loop over period and call the GT model repeatedly.
In models with individual feedback by time of day, they can
call this macro for just the current period. This matches the
DC model.