Improve/Correct disagg model
Previously, the disagg model assumed that all marginals would
use something like "average zone size". This doesn't work for
income, which uses a ratio of the zone's average/median to
the regional average/median. Now, the disagg models
assumes that the master index field already exists, and does
not try to create it internally.

As a result, fields like avg_size or inc_ratio are calculated
in the project code before calling the model.